+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Use the `org.realityforge.proton:proton-processor-pack` project for processor utility methods.
+
 ### [v0.21](https://github.com/realityforge/router-fu/tree/v0.21) (2019-12-17) · [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.20...v0.21)
 
 * Upgrade the `org.realityforge.arez` artifacts to version `0.159`.